src/backend/pyproject.toml
[project]
name = "registries-backend"
version = "0.1.0"
description = "Backend service for the monorepo"
requires-python = ">=3.14,<3.15"
dependencies = [
    "alembic>=1.16.0,<2.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "fastapi>=0.122.0,<1.0.0",
    "greenlet>=3.1.1,<4.0.0",
    "opentelemetry-exporter-otlp>=1.38.0,<2.0.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0,<1.0.0",
    "opentelemetry-instrumentation-redis>=0.59b0,<1.0.0",
    "opentelemetry-instrumentation-sqlalchemy>=0.59b0,<1.0.0",
    "opentelemetry-sdk>=1.38.0,<2.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.11.0,<3.0.0",
    "prometheus-fastapi-instrumentator>=7.1.0,<8.0.0",
    "redis>=5.2.0,<8.0.0",
    "sentry-sdk[fastapi]>=2.29.0,<3.0.0",
    "sqlalchemy>=2.0.30,<3.0.0",
    "starlette>=0.52.1,<1.0.0",
    "structlog>=25.4.0,<26.0.0",
    "taskiq>=0.11.18,<1.0.0",
    "taskiq-redis>=1.0.0,<2.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6,<2.0.0",
    "deptry>=0.23.1,<1.0.0",
    "eradicate>=3.0.0,<4.0.0",
    "httpx>=0.27.0,<1.0.0",
    "import-linter>=2.11,<3.0.0",
    "mypy>=1.18.2,<2.0.0",
    "pip-audit>=2.9.0,<3.0.0",
    "pre-commit>=4.2.0,<5.0.0",
    "pytest>=8.3.0,<10.0.0",
    "pytest-asyncio>=0.24.0,<2.0.0",
    "pyupgrade>=3.21.0,<4.0.0",
    "ruff>=0.12.0,<1.0.0",
    "tryceratops>=2.4.1,<3.0.0",
    "xenon>=0.9.3,<1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py314"
line-length = 120
src = ["api", "apps", "core", "runtime", "tests"]

[tool.ruff.lint]
select = ["B", "C4", "E9", "F811", "F821", "F822", "F823", "I", "PERF", "RET", "RUF", "SIM", "T20", "UP"]
ignore = ["B008", "RUF012"]

[tool.mypy]
python_version = "3.14"
strict = true
ignore_missing_imports = true
explicit_package_bases = true
namespace_packages = true
show_error_codes = true
pretty = true
warn_unused_configs = true

[tool.importlinter]
root_packages = ["api", "apps", "core", "runtime"]
exclude_type_checking_imports = true

[tool.deptry]
known_first_party = ["api", "apps", "core", "runtime"]

[[tool.importlinter.contracts]]
name = "Enforce backend top-level package boundaries"
type = "layers"
layers = ["api", "runtime", "apps", "core"]
ignore_imports = [
    "core.bootstrap.* -> api",
    "core.bootstrap.app -> apps.system.contracts",
    "core.bootstrap.system -> apps.system.application.services",
    "core.bootstrap.system -> apps.system.infrastructure.repositories",
]

[tool.uv]
package = false
